Make everything single threaded!
If content is ready from the buffer pipe, we read as much as we can into
a limitless memstream buffer. The tokenizer then pulls from that buffer
instead of directly from the pipe. This is to prevent the pipe from
filling up if the tokenizer is running slowly.

To support many annotators stopping and starting as they reach the end
of the buffer and more content is added, we dynamically build a `struct
pollfd` for `poll`. Write ends of annotators will only be included in
the poll if they are not already at the end of their buffer.